// finals-2024-01-16/code/5048-dijana-damcevska/src/app/models/band.ts
export interface Band {
    id: number;
    name: string;
    genre: string;
    formed: number;
    location: string;
    members: string[];
    albums: Album[];
}

export interface Album {
    name: string;
    year: number;
}

export interface Place {
    id: number,
    country: string,
    city: string[]
}